smooth scaling, auto-desaturation of player-colors for graph
