--- Assets/DistributionGraph.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 492ce39c9926f45d8a9b3ddc1cbf990b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
