--- Assets/Scripts/GUI/CalculateSwarmSizes.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f93877e459441bf4abb927a3ec01895b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
