+++ Assets/Scripts/GUI/DistributionGraph/DisplaySegment.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5304a49212a8e4f46b8364534372fcc8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
